// src/types/production/execution/processSubmissionList.ts
export interface ProcessSubmissionList {
  /**
   * 序号
   * 类型：数字(0.) 属性：只读
   */
  id: number;

  /**
   * 交接班日志编号
   * 类型：字符(50) 属性：只读
   * 详细规则与说明：
   * 再造烟叶线：再造烟叶前处理工序｜再造烟叶原料投用｜再造烟叶提取工序｜再造烟叶浓缩段CIP站｜再造烟叶浓缩工序｜再造烟叶配料工序｜再造烟叶配料间香料｜再造烟叶制浆一楼｜再造烟叶制浆二楼｜再造烟叶制浆DCS｜再造烟叶抄造湿部现场｜再造烟叶抄造湿部DCS｜再造烟叶抄造干部
   * 再造烟叶成品打包｜再造烟叶成品烘干｜再造烟叶浆液平衡
   * 再造梗丝线：再造梗丝原料预处理段｜再造梗丝润压梗段｜再造梗丝切丝段｜再造梗丝浸泡提取段｜再造梗丝提取后烘干段｜再造梗丝回填加料段｜再造梗丝回填后烘丝段｜再造梗丝贮梗丝段｜再造梗丝净化段｜再造梗丝浓缩段｜再造梗丝除尘段｜再造梗丝回填料液配制段
   * 香精香料线：香精香料线
   */
  logNo: string;

  /**
   * 是否提交
   * 类型：布尔 属性：只读
   * 详细规则与说明：有效值：已提交｜未提交。规则：对应工序的操作工交接班日志是否提交
   */
  isSubmitted: boolean;

  /**
   * 提交人
   * 类型：字符(20) 属性：只读
   * 详细规则与说明：关联：操作工-交接班日志-提交人
   */
  submitter?: string;

  /**
   * 提交时间
   * 类型：时间（yyyy-mm-dd hh:mm:ss） 属性：只读
   * 详细规则与说明：关联：操作工-交接班日志-提交时间
   */
  submitTime?: string;

  /**
   * 生产任务ID
   */
  taskId: number;
}